No need for custom writepanel.js if using the supported WooCommerce method for showing and hiding metabox tabs and options
WooCommerce already has the functionality for showing and hiding
meta-box tabs and options in meta-boxes-product.js. You simply need to
add the class "show_if_" + product-type or "hide_if_" + product-type.
Our product-type is "accommodation-booking" (note the WooCommerce uses a
"-" for separate words in a product type, not an "_"). So by assigning
the classes "show_if_accommodation-booking" and
"hide_if_accommodation-booking", we can get rid of the unneeded
writepanel.js file and avoid the associated problems that came from this
hack.