Don't force has_resources() to return true
The WC_Product_Booking object that we're extending seems to have a
perfectly functional has_resources() function. There is no need to
override it here. Forcing this to return true causes a problem when
Accommodations may or may not have Resources depending on whether or not
the "Has resources" checkbox is selected in the admin.