Add WC_Accommodation_Booking_Addons class
This class is based on the WC_Bookings_Addons class from the Bookings
module. Unfortunately, there is no way to access the instance of that
class without modifying the Bookings plugin so it is not possible for us
to unhook or modify any of its actions or filters. However, we can
display our own set of options and allow them to save the multipliers
appropriately.