Add get_blocks_in_range() to product class
The WC_Product_Booking classes' get_blocks_in_range() function does not
account for 'nights' as a valid duration unit. So it defaults to
returning every minute of every day as a block, resulting in 1439 blocks
instead of 1! This severly slows down the page load time if you have an
accommodations booking with multiple resources.