Fix availability calculation issue with is_booked_on_day function
With multiday bookings, the is_booked_on_day method of the WC_Booking
class only checks whether the block ends after the booking date starts
or whether the block starts before the booking date ends. This doesn't
take into account blocks that both start and end before the booking date
starts or after it ends.

It is very possible that a booking may not intersect at all with the
block being checked in the case of booking accommodations. The
get_blocks_availability function of the WC_Product_Booking class that
calls this is_booked_on_day method, grabs all of the existing bookings
in a given range and then examines each block in that range
individually. Obviously, it's very likely that some of the bookings in
that range would fall outside of the block being examined.