Correct error returned by get_blocks_availability
Override the get_blocks_availability method of the WC_Product_Booking
class to correct the error message. It now accurately gives the
available quantity of places remaining on the day that triggered the
error instead of giving the maximum number of available places open on
one of the days in the range.