Accomodation duration unit is always one night.
Accommodation differs from the standard booking product as the duration is always one day.
Overwriting this function acts as a compatibility layer. It is possible to create a and set
paramaters to a booking prodcut. Later it is possible to change the product type to accommodation.
Some of the fields are hidden but still remain set. This prevents now
widden duration field from havving an effect on the calculation.