Add a 'NumberUtil' class with a 'abs' method, and PSR4 infrastructure
There's a number of places in the code where the built-in function
'abs' is executed passing a non-numeric value (not a number and not
a string that can be parsed as a number), for example abs('').
In PHP 7 this yields a value of 0, but in PHP 8 this throws an error.

This commit adds a 'NumberUtil' class with a static 'abs' method,
this method applies 'floatval' to the supplied value and thus if
it's not numeric it will be interpreted as 0, as PHP 7 does; then
it just executes the built-in function. And all the calls to 'abs' in
the codebase are replaced with 'NumberUtil::abs'.

Additionally, infrastructure has been added so that new code can go
into a 'src' folder following PSR4 namespacing rules (and similarly
for unit tests).